Fix sphinx Zeitwerk problems
Sphinx indices must be under app/indices and out of the autoload path.
Modules like AccountIndex must be in the autoload path.
