+++ app/models/topic.rb
@@ -7,7 +7,7 @@ class Topic < ApplicationRecord
   include ThreeScale::SpamProtection::Integration::Model
   has_spam_protection
 
-  include TopicIndex
+  include Indices::TopicIndex
 
   ORDER_BY = [['Newest post', ''], ["Most views", "hits"], ["Number of posts", "posts_count"]]
