+++ test/unit/indices/indexing_test.rb
@@ -57,8 +57,8 @@ class IndexingTest < ActiveSupport::TestCase
   # the idea is to double check #indexed_models lists all models, assuring we cover them all in tests
   test "all models with index methods are indexed" do
     exclusions = [ActiveStorage::Blob, ActiveStorage::Attachment, ApplicationRecord, Plan, Cinstance, User]
-    index_modules = [Searchable, AccountIndex::ForAccount]
-    index_modules << TopicIndex unless System::Database.oracle?
+    index_modules = [Searchable, Indices::AccountIndex::ForAccount]
+    index_modules << Indices::TopicIndex unless System::Database.oracle?
 
     models = ActiveRecord::Base.descendants.select do |model|
       index_modules.any? { |mod| mod === model.new } unless exclusions.include?(model)
